Extract out task runner
